arxiv_chatbot/requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
kaggle
scikit-learn
numpy
pandas
